third_party/blink/web_tests/external/wpt/css/css-variables/variable-cssText-expected.txt

This is a testharness.js-based test.
PASS target1
PASS target2
PASS target3
FAIL target4 assert_equals: expected "margin: var(--prop) !important;" but got "margin: var(--prop)  !important;"
FAIL target5 assert_equals: expected "background: var(--prop) !important;" but got "background: var(--prop)  !important;"
PASS target6
PASS target7
PASS target8
PASS target9
PASS target10
Harness: the test ran to completion.
